Extract LeaveRequestsTimelineTransformer
